Scrolling now works properly in SMB
Due to previously running the CPU and PPU for one scanline each, mid scanline writes to the scroll register were not handled properly. Rather than change the PPU to accommodate this, the scanline now runs in two different steps - one for the visible pixels, and one for the end of the scanline. This way the scroll writes apply at the appropriate time rather than one scanline too late.